Replace WRS Mayan heiroglyph background w/ solid grey
Since I use the WRS Mayan heiroglyph as my desktop background, it looks
weird having it w/in my browser on top of my desktop.  It may also be too
busy for general use on the site.  Replace it w/ a solid grey background
for now.  Consider other possibilities later.